# docs/swagger.yaml
basePath: /
definitions:
...
  controller._ResponseRoleList:
    properties:
      code:
        allOf:
        - $ref: '#/definitions/controller.ResCode'
        description: 业务响应状态码
      data:
        $ref: '#/definitions/model.ResponseRoleList'
      message:
        description: 提示信息
        type: string
      ok:
        type: boolean
    type: object
...
  controller.ResCode:
    enum:
    - 200
    - 201
    - 202
    - 203
    - 204
    - 205
    - 206
    - 207
    - 208
    type: integer
    x-enum-varnames:
    - CodeSuccess
    - CodeInvalidParam
    - CodeUserExist
    - CodeUserNotExist
    - CodeInvalidPassword
    - CodeServerBusy
    - CodeInvalidToken
    - CodeNeedLogin
    - CodeMenuNodeExist
  controller.ResponseData:
    properties:
      code:
        $ref: '#/definitions/controller.ResCode'
      data: {}
      message: {}
      ok:
        type: boolean
    type: object
...
  model.ParamRoleSave:
    properties:
      remark:
        description: 角色备注
        type: string
      roleName:
        description: 角色名称，不能为空
        type: string
    type: object
  model.ParamRoleUpdate:
    properties:
      remark:
        description: 角色备注
        type: string
      role_id:
        description: 用户ID，
        type: integer
      roleName:
        description: 角色名称，不能为空
        type: string
    type: object
  model.ParamTmSave:
    properties:
      logoUrl:
        type: string
      tmName:
        type: string
    required:
    - logoUrl
    - tmName
    type: object
  model.ParamTmUpdate:
    properties:
      id:
        type: integer
      logoUrl:
        type: string
      tmName:
        type: string
    required:
    - logoUrl
    - tmName
    type: object
...
  model.ResponseRoleList:
    properties:
      current:
        type: integer
      pages:
        type: integer
      records:
        items:
          $ref: '#/definitions/model.Role'
        type: array
      searchCount:
        type: boolean
      size:
        type: integer
      total:
        type: integer
    type: object
  model.ResponseTmList:
    properties:
      current:
        type: integer
      pages:
        type: integer
      records:
        items:
          $ref: '#/definitions/model.Trademark'
        type: array
      searchCount:
        type: boolean
      size:
        type: integer
      total:
        type: integer
    type: object
...
  model.Role:
    properties:
      ID:
        type: integer
      createTime:
        type: string
      id:
        type: integer
      remark:
        type: string
      roleName:
        type: string
      updateTime:
        type: string
    type: object
  model.Trademark:
    properties:
      ID:
        type: integer
      createTime:
        type: string
      id:
        type: integer
      logoUrl:
        type: string
      tmName:
        type: string
      updateTime:
        type: string
    type: object
host: 127.0.0.1:10086
info:
  contact:
    email: 929000201@qq.com
    name: WangJian
    url: https://github.com/wangjian3306
  description: 硅谷甄选项目后端
  title: vue3_admin 项目接口文档
  version: "1.0"
paths:
...
  /admin/acl/role/{page}/{limit}:
    get:
      consumes:
      - application/json
      description: 获取角色列表
      parameters:
      - description: 用户 Token
        in: header
        name: token
        required: true
        type: string
      - description: 当前页码
        in: path
        name: page
        required: true
        type: integer
      - description: 每页记录数
        in: path
        name: limit
        required: true
        type: integer
      - description: 角色名
        in: query
        name: roleName
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller._ResponseRoleList'
      security:
      - ApiKeyAuth: []
      summary: 获取角色分页列表
      tags:
      - 角色管理
  /admin/acl/role/remove/{id}:
    delete:
      consumes:
      - application/json
      description: 处理删除角色请求
      parameters:
      - description: 用户 Token
        in: header
        name: token
        required: true
        type: string
      - description: 角色 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: 删除角色接口
      tags:
      - 角色管理
  /admin/acl/role/save:
    post:
      consumes:
      - application/json
      description: 处理新增角色请求
      parameters:
      - description: 用户 Token
        in: header
        name: token
        required: true
        type: string
      - description: 用户登录参数
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/model.ParamRoleSave'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: 新增角色接口
      tags:
      - 角色管理
  /admin/acl/role/update:
    put:
      consumes:
      - application/json
      description: 处理更新角色请求
      parameters:
      - description: 用户 Token
        in: header
        name: token
        required: true
        type: string
      - description: 用户登录参数
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/model.ParamRoleUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: 更新角色接口
      tags:
      - 角色管理
...
  /admin/product/baseTrademark/{page}/{limit}:
    get:
      consumes:
      - application/json
      description: 获取品牌列表
      parameters:
      - description: 用户 Token
        in: header
        name: token
        required: true
        type: string
      - description: 当前页码
        in: path
        name: page
        required: true
        type: integer
      - description: 每页记录数
        in: path
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ResponseTmList'
      security:
      - ApiKeyAuth: []
      summary: 获取品牌分页列表
      tags:
      - 品牌管理
  /admin/product/baseTrademark/remove/{id}:
    delete:
      consumes:
      - application/json
      description: 处理删除品牌请求
      parameters:
      - description: 用户 Token
        in: header
        name: token
        required: true
        type: string
      - description: 品牌 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: 删除品牌接口
      tags:
      - 品牌管理
  /admin/product/baseTrademark/save:
    post:
      consumes:
      - application/json
      description: 新增品牌接口
      parameters:
      - description: 用户 Token
        in: header
        name: token
        required: true
        type: string
      - description: 用户角色
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/model.ParamTmSave'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: 新增品牌接口
      tags:
      - 品牌管理
  /admin/product/baseTrademark/update:
    put:
      consumes:
      - application/json
      description: 处理更新品牌请求
      parameters:
      - description: 用户 Token
        in: header
        name: token
        required: true
        type: string
      - description: 用户登录参数
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/model.ParamTmUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.ResponseData'
      security:
      - ApiKeyAuth: []
      summary: 更新品牌接口
      tags:
      - 品牌管理
...
swagger: "2.0"